minimal change reorge before further work
